+++ packages/twenty-front/src/modules/action-menu/actions/record-actions/components/RegisterRecordActionEffect.tsx
@@ -1,4 +1,4 @@
-import { RecordConfigAction } from '@/action-menu/actions/types/ConfigAction';
+import { RecordConfigAction } from '@/action-menu/actions/types/RecordConfigAction';
 import { wrapActionInCallbacks } from '@/action-menu/actions/utils/wrapActionInCallbacks';
 import { ActionMenuContext } from '@/action-menu/contexts/ActionMenuContext';
 import { useActionMenuEntries } from '@/action-menu/hooks/useActionMenuEntries';
