+++ packages/twenty-front/src/modules/action-menu/actions/types/RecordAgnosticConfigAction.ts
@@ -1,0 +1,9 @@
+import { ActionHookWithoutObjectMetadataItem } from '@/action-menu/actions/types/ActionHook';
+import { RecordConfigAction } from '@/action-menu/actions/types/RecordConfigAction';
+
+export type RecordAgnosticConfigAction = Omit<
+  RecordConfigAction,
+  'useAction'
+> & {
+  useAction: ActionHookWithoutObjectMetadataItem;
+};
